requirements.txt
# Core Deep Learning Libraries
torch>=2.0.0
torchvision>=0.15.0
tensorflow>=2.13.0

# Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Model Explainability
grad-cam>=1.4.0

# Web Deployment
streamlit>=1.28.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0

# Jupyter Notebooks
jupyter>=1.0.0
ipykernel>=6.25.0
notebook>=7.0.0

# Utilities
tqdm>=4.66.0



